src/requirements.txt
tensorflow==1.13.1
numpy==1.19.2
pillow==8.0.1
word2vec==0.11.1